api/requirements.txt
Django==4.2.0
python-decouple==3.8
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
opensearch-py==2.4.0
aiohttp
requests
google-genai
gunicorn==21.2.0